requirements.txt
alembic==1.13.0
APScheduler==3.6.3
blinker==1.7.0
cachetools==4.2.2
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
Flask==3.0.0
greenlet==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
python-telegram-bot==13.13
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
tornado==6.4
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
Werkzeug==3.0.1
